web: inject version-with-commit via Vite define; use in UI; remove SvelteKit kit.version usage
