+++ packages/web/src/types/app-version.d.ts
@@ -1,0 +1,1 @@
+declare const __APP_VERSION__: string
